Added: Support for e621 (module) Bugs Changes: Still supporting derpibooru (module now) Removed: Some code
