+++ dermod/follow.py
@@ -4,7 +4,7 @@ import settings_file
 import sys
 import os
 import shutil
-from dermod import db, derpilist_v2, derpiload_v3
+from dermod import db, listloader, imgloader
 
 
 def run(run_once=False):
@@ -14,8 +14,10 @@ def run(run_once=False):
     while True:
         print("Checking new images")
         try:
-            derpilist_v2.run(follower=True, pages_num=settings_file.checked_pages)
-            derpiload_v3.run(file=settings_file.ids_file, check_local=False)
+            for i in settings_file.modules:
+            module = importlib.import_module('dermod.sitesupport.{}'.format(i))
+            listloader.run(follower=True, pages_num=settings_file.checked_pages, module=module)
+            imgloader.run(file=settings_file.ids_file, check_local=False)
             db.fill_db(file=settings_file.ids_file)
         except Exception as exc:
             print("Exception occured: "+exc.with_traceback(None))
